requirements.txt
beautifulsoup4==4.12.3
blinker==1.8.2
Bootstrap-Flask==2.4.0
bs4==0.0.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
environs==11.0.0
fake-useragent==1.5.1
Flask==3.0.3
Flask-Limiter==3.5.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
marshmallow==3.21.3
packaging==24.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
soupsieve==2.5
timeago==1.0.16
urllib3==2.2.2
Werkzeug==3.0.3
WTForms==3.1.2
